+++ scripts/ard/categories.js
@@ -3,7 +3,7 @@ const undici = require('../../packages/undici')()
 const createHashedId = require('../../utils/ard/createHashedId')
 
 // init storage
-const Storage = require('../../packages/storage-wrapper')
+const Storage = require('../../packages/storage')
 
 const storage = new Storage({
 	gs: { projectId: process.env.GCP_PROJECT_ID },