+++ scripts/ard/sortPubByExtId.js
@@ -1,13 +1,11 @@
 /*
 
-	by SWR audio lab
-
 	sorts list of ard-publishers by external-id
 
 */
 
 // init storage
-const Storage = require('../../packages/storage-wrapper')
+const Storage = require('../../packages/storage')
 
 const storage = new Storage({
 	gs: { projectId: process.env.GCP_PROJECT_ID },