+++ utils/storage/createUri.js
@@ -1,11 +1,3 @@
-/*
-
-	node-storage-wrapper
-
-	This module provides easy access to combine bucket + path to unique URIs
-
-*/
-
 module.exports = {
 	s3: (bucket, path) => {
 		// return data