+++ utils/storage/save.js
@@ -1,14 +1,10 @@
 /* eslint-disable func-names */
-/*
-
-	node-storage-wrapper
-
-*/
 
 // load node utils
 const os = require('os')
 const pathUtil = require('path')
 const { v4: uuidv4 } = require('uuid')
+const { PutObjectCommand } = require('@aws-sdk/client-s3')
 
 const saveLocalFile = (that, uri, buffer) =>
 	new Promise((resolve, reject) => {
@@ -26,44 +22,55 @@ const deleteLocalFile = (that, filePath) =>
 		})
 	})
 
-module.exports = async function (uri, buffer, logPrefix, resumable) {
-	const thisLogPrefix = logPrefix ? [logPrefix, '>'] : []
-	let structure, bucket, path
-
+module.exports = async function (uri, buffer, resumable) {
 	if (uri.substr(0, 5).toLowerCase() === 's3://') {
 		// aws s3 file
-		structure = uri.substr(5).split('/')
-		bucket = structure.shift()
-		path = structure.join('/')
+		const structure = uri.substr(5).split('/')
+		const bucket = structure.shift()
+		const path = structure.join('/')
 
 		// log progress
-		this.sdk.log(this, 'log', thisLogPrefix.concat(['storage.save.s3 >', uri]))
+		if (this.logger) {
+			this.logger.log({
+				level: 'info',
+				message: `storage.save.s3 > ${uri}`,
+				source: this.logSource,
+				data: { uri },
+			})
+		}
 
 		// upload to aws
-		await this.sdk.s3
-			.upload({
+		const saved = await this.sdk.s3.send(
+			new PutObjectCommand({
 				Bucket: bucket,
 				Body: buffer,
 				Key: path,
 			})
-			.promise()
+		)
 
 		// return ok
-		return Promise.resolve()
+		return Promise.resolve(saved)
 	}
 
 	if (uri.substr(0, 5).toLowerCase() === 'gs://') {
 		// google cloud storage
-		structure = uri.substr(5).split('/')
-		bucket = structure.shift()
-		path = structure.join('/')
+		const structure = uri.substr(5).split('/')
+		const bucket = structure.shift()
+		const path = structure.join('/')
 
 		// save to local file
 		const tempFilePath = pathUtil.resolve(os.tmpdir(), uuidv4())
 		await saveLocalFile(this, tempFilePath, buffer)
 
 		// log progress
-		this.sdk.log(this, 'log', thisLogPrefix.concat(['storage.save.gs >', uri]))
+		if (this.logger) {
+			this.logger.log({
+				level: 'info',
+				message: `storage.save.gs > ${uri}`,
+				source: this.logSource,
+				data: { uri, resumable },
+			})
+		}
 
 		// create default bucket config
 		const bucketConfig = {
@@ -78,19 +85,26 @@ module.exports = async function (uri, buffer, logPrefix, resumable) {
 		}
 
 		// upload file to gcs
-		await this.sdk.gs.bucket(bucket).upload(tempFilePath, bucketConfig)
+		const [saved] = await this.sdk.gs.bucket(bucket).upload(tempFilePath, bucketConfig)
 
 		// delete local temp file
 		await deleteLocalFile(this, tempFilePath)
 
 		// return ok
-		return Promise.resolve()
+		return Promise.resolve(saved)
 	}
 
 	// local file
 
 	// log progress
-	this.sdk.log(this, 'log', thisLogPrefix.concat(['storage.save.local >', uri]))
+	if (this.logger) {
+		this.logger.log({
+			level: 'info',
+			message: `storage.save.local > ${uri}`,
+			source: this.logSource,
+			data: { uri },
+		})
+	}
 
 	// save file
 	const file = await saveLocalFile(this, uri, buffer)